Add stat app type to FlowTag and move FlowCard to component folder
